requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
dnspython==2.7.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
pymongo==4.10.1
python-telegram-bot==21.6
sniffio==1.3.1
typing_extensions==4.12.2
